Change BUG_ON k2_rb_root.
Do not Oops, if the idx is out of range. Instead clamp it to idle
priority and issue a rate-limited warning. Same goes for k2_queue.

Only Oops if data direction is neither read nor write. There are ways to
avoid Oopsing in those cases too, but data direction should be correct.